refactor: change transaction modal name
